Checking if a procress has exited before closing it, in login

+++ MPAi/NewForms/LoginScreen.cs
@@ -271,7 +271,10 @@ namespace MPAi.NewForms
             if (!(PlotController.getCurrentPlotProcess() == null))
             {
                 //Kills any persistent Procecsses.
-                PlotController.getCurrentPlotProcess().Kill();
+                if (!(PlotController.getCurrentPlotProcess().HasExited))
+                {
+                    PlotController.getCurrentPlotProcess().Kill();
+                }
             }
 
             UserManagement.WriteSettings();
